feat(profileCtrl.ts, profileRoutes.ts, AuthProvider.tsx, apiProvider.ts, Caroussel.tsx, User.type.ts): add functionality to delete a picture from user's profile
Add a new endpoint to delete a picture from user's profile. Implement the functionality in the profileCtrl.ts file and add the corresponding route in profileRoutes.ts. In the client-side, add a new function to delete a picture in the AuthProvider.tsx file and implement it in the Caroussel.tsx component. Also, add a new method to the apiProvider.ts file to make the request to the backend. Finally, add the id field to the UserProfile type in the User.type.ts file.